feat: fix space permission + e2e tests

+++ packages/backend/src/services/SavedChartsService/SavedChartService.ts
@@ -237,7 +237,7 @@ export class SavedChartService {
         user: SessionUser,
         savedChartUuid: string,
     ): Promise<SavedChart> {
-        const { organizationUuid, projectUuid, pinnedListUuid } =
+        const { organizationUuid, projectUuid, pinnedListUuid, spaceUuid } =
             await this.savedChartModel.getSummary(savedChartUuid);
 
         if (
@@ -248,6 +248,11 @@ export class SavedChartService {
         ) {
             throw new ForbiddenError();
         }
+
+        if (!(await this.hasChartSpaceAccess(spaceUuid, user.userUuid))) {
+            throw new ForbiddenError();
+        }
+
         if (pinnedListUuid) {
             await this.pinnedListModel.deleteItem({
                 pinnedListUuid,
@@ -259,7 +264,6 @@ export class SavedChartService {
                 savedChartUuid,
             });
         }
-
         const pinnedList = await this.pinnedListModel.getPinnedListAndItems(
             projectUuid,
         );
@@ -297,6 +301,16 @@ export class SavedChartService {
         ) {
             throw new ForbiddenError();
         }
+
+        const spaceAccessPromises = data.map(async (chart) =>
+            this.hasChartSpaceAccess(chart.spaceUuid, user.userUuid),
+        );
+
+        const hasAllAccess = await Promise.all(spaceAccessPromises);
+        if (hasAllAccess.includes(false)) {
+            throw new ForbiddenError();
+        }
+
         const savedCharts = await this.savedChartModel.updateMultiple(data);
         analytics.track({
             event: 'saved_chart.updated_multiple',
@@ -366,6 +380,7 @@ export class SavedChartService {
         if (user.ability.cannot('view', subject('SavedChart', savedChart))) {
             throw new ForbiddenError();
         }
+
         if (
             !(await this.hasChartSpaceAccess(
                 savedChart.spaceUuid,
@@ -409,6 +424,15 @@ export class SavedChartService {
         ) {
             throw new ForbiddenError();
         }
+        if (savedChart.spaceUuid) {
+            const space = await this.spaceModel.getFullSpace(
+                savedChart.spaceUuid,
+            );
+            if (!hasSpaceAccess(space, user.userUuid)) {
+                throw new ForbiddenError();
+            }
+        }
+
         const newSavedChart = await this.savedChartModel.create(
             projectUuid,
             user.userUuid,
